increase timeout, fix data races, and refactor 2 tests into their own class
